viewmodel setup, adapter setup and view integration with data

+++ app/src/main/java/com/samples/bitclassassignment/domain/Model.kt
@@ -1,5 +1,6 @@
 package com.samples.bitclassassignment.domain
 
+
 /**
  * Created by ak93.droid@gmail.com on 23,August,2020
  */
@@ -7,5 +8,5 @@ data class LessonDetails(val id : Int,
                           val createdAt : String,
                           val code: String,
                           val heading: String,
-                          val startTime : Int,
-                          val endTime : Int)
+                          val startTime : Long,
+                          val endTime : Long)
